pyproject.toml
[project]
name = "fm-comfyui-bridge"
version = "0.10.0"
description = "ComfyUI access bridge"
readme = "README.md"
authors = [
    { name = "rerofumi", email = "rero2@yuumu.org" }
]
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "Topic :: Software Development :: Libraries :: Python Modules",
]
requires-python = ">=3.10"
dependencies = [
    "pillow>=11.0.0",
    "pyyaml>=6.0.2",
    "requests>=2.32.3",
    "httpx>=0.28.1",
    "websockets>=15.0.1",
]

[project.urls]
Homepage = "https://github.com/rerofumi/fm_comfyui_bridge"

[project.scripts]
fm-comfy-request = "fm_comfy_request.cli:main"

[tool.hatch.build.targets.sdist]
exclude = [
    "/.jj"
]


[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.pytest.ini_options]
minversion = "6.0"
addopts = "-ra -q"
testpaths = [
    "tests",
]

[dependency-groups]
dev = [
    "pytest>=8.3.3",
]


